examples: refactor headers
