+++ examples/bpftools/bindsnoop/bindsnoop.bpf.c
@@ -5,7 +5,7 @@
 #include <bpf/bpf_core_read.h>
 #include <bpf/bpf_tracing.h>
 #include <bpf/bpf_endian.h>
-#include "bindsnoop.bpf.h"
+#include "bindsnoop.h"
 
 #define MAX_ENTRIES	10240
 #define MAX_PORTS	1024
