+++ examples/bpftools/bootstrap/bootstrap.bpf.c
@@ -4,7 +4,7 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
 #include <bpf/bpf_core_read.h>
-#include "bootstrap.bpf.h"
+#include "bootstrap.h"
 
 char LICENSE[] SEC("license") = "Dual BSD/GPL";
