+++ examples/bpftools/mdflush/mdflush.bpf.c
@@ -5,7 +5,7 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_core_read.h>
 #include "core_fixes.h"
-#include "mdflush.bpf.h"
+#include "mdflush.h"
 
 struct {
 	__uint(type, BPF_MAP_TYPE_PERF_EVENT_ARRAY);
