+++ examples/bpftools/tcpstates/tcpstates.bpf.c
@@ -4,7 +4,7 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
 #include <bpf/bpf_core_read.h>
-#include "tcpstates.bpf.h"
+#include "tcpstates.h"
 
 #define MAX_ENTRIES 10240
 #define AF_INET 2
